ao_alsa: remove some dead code
This was only added recently (74234e17) as an attempt to minimize the
bad impact of channel layout device aliases. But use of these was
removed in commit d690c055. Now this code does pretty much nothing, and
shouldn't be needed anymore. It does something when using spdif, but
this fallback won't work anyway.